[2.5.1] update to 2.5.1

+++ format.sh
@@ -1,6 +1,8 @@
 #!/bin/bash
 
-for f in assets/minecraft/lang/*.json; do
-    raw="$(sed -E 's,(\\u(e|f[78])),\\\1,g' "$f")"
-    echo "$raw" | jq --sort-keys | sed 's|\\\\|\\|g' > "$f"
+for f in assets/minecraft/{lang,lang_rpo}/*.{json,rpo}; do
+    if [ -f "$f" ]; then
+        raw="$(sed -E 's,(\\u(e|f[78])),\\\1,g' "$f")"
+        echo "$raw" | jq --sort-keys | sed 's|\\\\|\\|g' > "$f"
+    fi
 done
